+++ src/app/(BANK_PAGES)/components/Footer.tsx
