+++ src/app/(BANK_PAGES)/components/smallFooter.tsx
