+++ src/app/(DASHBOARD_PAGES)/components/dashboard/MonthlyEarnings.tsx
