+++ src/app/(DASHBOARD_PAGES)/components/dashboard/ProductPerformance.tsx
