+++ src/app/(DASHBOARD_PAGES)/components/dashboard/UserRegistration.tsx
