+++ src/app/(DASHBOARD_PAGES)/components/shared/BlankCard.tsx
